backend/requirements.txt
fastapi
uvicorn[standard]
pyjwt
pydantic
databases[sqlite]
aiosqlite